Move GenericManifestEntry to a separate file.
