+++ core/src/main/java/org/apache/iceberg/ManifestEntry.java
@@ -81,137 +81,3 @@ interface ManifestEntry {
 
   ManifestEntry copyWithoutStats();
 }
-
-class GenericManifestEntry implements ManifestEntry, IndexedRecord, SpecificData.SchemaConstructable {
-  private final org.apache.avro.Schema schema;
-  private final IndexedDataFile fileWrapper;
-  private Status status = Status.EXISTING;
-  private Long snapshotId = null;
-  private DataFile file = null;
-
-  GenericManifestEntry(org.apache.avro.Schema schema) {
-    this.schema = schema;
-    this.fileWrapper = null; // do not use the file wrapper to read
-  }
-
-  GenericManifestEntry(StructType partitionType) {
-    this.schema = AvroSchemaUtil.convert(V1Metadata.entrySchema(partitionType), "manifest_entry");
-    this.fileWrapper = new IndexedDataFile(schema.getField("data_file").schema());
-  }
-
-  private GenericManifestEntry(GenericManifestEntry toCopy, boolean fullCopy) {
-    this.schema = toCopy.schema;
-    this.fileWrapper = new IndexedDataFile(schema.getField("data_file").schema());
-    this.status = toCopy.status;
-    this.snapshotId = toCopy.snapshotId;
-    if (fullCopy) {
-      this.file = toCopy.file().copy();
-    } else {
-      this.file = toCopy.file().copyWithoutStats();
-    }
-  }
-
-  ManifestEntry wrapExisting(Long newSnapshotId, DataFile newFile) {
-    this.status = Status.EXISTING;
-    this.snapshotId = newSnapshotId;
-    this.file = newFile;
-    return this;
-  }
-
-  ManifestEntry wrapAppend(Long newSnapshotId, DataFile newFile) {
-    this.status = Status.ADDED;
-    this.snapshotId = newSnapshotId;
-    this.file = newFile;
-    return this;
-  }
-
-  ManifestEntry wrapDelete(Long newSnapshotId, DataFile newFile) {
-    this.status = Status.DELETED;
-    this.snapshotId = newSnapshotId;
-    this.file = newFile;
-    return this;
-  }
-
-  /**
-   * @return the status of the file, whether EXISTING, ADDED, or DELETED
-   */
-  public Status status() {
-    return status;
-  }
-
-  /**
-   * @return id of the snapshot in which the file was added to the table
-   */
-  public Long snapshotId() {
-    return snapshotId;
-  }
-
-  /**
-   * @return a file
-   */
-  public DataFile file() {
-    return file;
-  }
-
-  public ManifestEntry copy() {
-    return new GenericManifestEntry(this, true /* full copy */);
-  }
-
-  public ManifestEntry copyWithoutStats() {
-    return new GenericManifestEntry(this, false /* drop stats */);
-  }
-
-  @Override
-  public void setSnapshotId(long snapshotId) {
-    this.snapshotId = snapshotId;
-  }
-
-  @Override
-  public void put(int i, Object v) {
-    switch (i) {
-      case 0:
-        this.status = Status.values()[(Integer) v];
-        return;
-      case 1:
-        this.snapshotId = (Long) v;
-        return;
-      case 2:
-        this.file = (DataFile) v;
-        return;
-      default:
-        // ignore the object, it must be from a newer version of the format
-    }
-  }
-
-  @Override
-  public Object get(int i) {
-    switch (i) {
-      case 0:
-        return status.id();
-      case 1:
-        return snapshotId;
-      case 2:
-        if (fileWrapper == null || file instanceof GenericDataFile) {
-          return file;
-        } else {
-          return fileWrapper.wrap(file);
-        }
-      default:
-        throw new UnsupportedOperationException("Unknown field ordinal: " + i);
-    }
-  }
-
-  @Override
-  public org.apache.avro.Schema getSchema() {
-    return schema;
-  }
-
-  @Override
-  public String toString() {
-    return MoreObjects.toStringHelper(this)
-        .add("status", status)
-        .add("snapshot_id", snapshotId)
-        .add("file", file)
-        .toString();
-  }
-}
